+++ src/App.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
-import Landing from './pages/Landing';
-import List from './pages/List';
+import Landing from './pages/Landing/Landing';
+import List from './pages/List/List';
 import Mypage from './pages/Mypage/Mypage';
 
 function App() {
