+++ src/pages/List/List.jsx
@@ -1,8 +1,7 @@
 import React from 'react';
-import Header from '../components/Header';
+import Header from '../../components/Header';
 import DonationsList from './components/DonationList';
-import ThisMonthsChart from '../components/ThisMonthsChart';
-
+import ThisMonthsChart from '../../components/ThisMonthsChart';
 function List() {
   return (
     <>
