+++ src/pages/List/components/DonationList.css
@@ -1,3 +1,5 @@
+@import '../../../components/Slide.css';
+
 .donation {
   width: 100%;
   max-width: 1200px;
@@ -109,6 +111,6 @@
 .icon-credit {
   width: 14px;
   height: 14px;
-  background: url(../../assets/images/ico_credit.png) no-repeat center center;
+  background: url(../../../assets/images/ico_credit.png) no-repeat center center;
   background-size: cover;
 }
